boards/hifive1: Enable SPI driver
Enable the SPI driver on the HiFive 1

This makes the following configurations choices for the HiFive 1:

The SPI 0 peripheral driver is not enabled by default because it is in
charge of mapping the SPI flash into memory. This can be configured
using the CONFIG_SIFIVE_SPI_0_ROM KConfig option.

The SPI 1 peripheral driver is enabled by default and the pinmux is
configured for all of its outputs

The SPI 2 peripheral driver is enabled by default because it is present
in the DTS for the FE310, but because the QFN48 package used on the
HiFive 1 doesn't route those pins from the silicon die, the pinmux can't
enable the SPI 2 pins.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>